+++ src/main/java/haspiev/dev/hw_01/operations/processors/DepositAccountProcessor.java
@@ -23,7 +23,7 @@ public class DepositAccountProcessor implements OperationCommandProcessor {
     @Override
     public void processOperation() {
         System.out.println("Enter account id:");
-        int accounId = Integer.parseInt(scanner.nextLine());
+        Long accounId = Long.parseLong(scanner.nextLine());
         System.out.println("Enter amount to deposit:");
         int amountToDeposit = Integer.parseInt(scanner.nextLine());
         accountService.depositAccount(accounId, amountToDeposit);
